requirements.txt
# Core
crewai==0.28.0
crewai-tools==0.1.6
composio-core==0.3.20
composio-crewai==0.3.20

# OpenAI
openai==1.12.0

# Database
sqlite3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
pydantic==2.6.0

# Logging
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0